+++ apps/smtp/src/modules/smtp/services/html-to-text-compiler.ts
@@ -1,8 +1,8 @@
 import { convert } from "html-to-text";
 
-import { createLogger } from "../../logger";
+import { createLogger } from "../../../logger";
 import { fromThrowable, Result } from "neverthrow";
-import { BaseError } from "../../errors";
+import { BaseError } from "../../../errors";
 
 export interface IHtmlToTextCompiler {
   compile(html: string): Result<string, InstanceType<typeof BaseError>>;
